tests/wpt/web-platform-tests/html/dom/dynamic-markup-insertion/document-write/script_009.html

<!doctype html>
<title>document.write script that document.writes script</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var order = [];
t.step(function() {
  order.push(1);
  document.write("<script>order.push(2); document.write('<script>order.push(3); document.write(\"<script>order.push(4);</script\"+\">\"); order.push(5);</script' + '>'); order.push(6);</script" + ">");
  order.push(7);
});
</script>
<script>
t.step(function() {
  assert_array_equals(order, [1,2,3,4,5,6,7]);
});
t.done();
</script>
<div id="log"></div>
